requirements.txt
asgiref==3.6.0
django-allauth==0.54.0
django-simple-tags==0.6.0
django-nonrelated-inlines==0.2
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
dj-database-url==1.2.0
Django==4.1.7
django-mssql-backend==2.8.1
djangorestframework==3.14.0
djlint==1.19.16
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.7
mssql==1.0.1
pathspec==0.11.1
psycopg2-binary==2.9.5
pyodbc==4.0.35
pytz==2023.2
PyYAML==6.0
regex==2022.10.31
six==1.16.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
